refactor: translate
