+++ src/main/java/com/chalim/backend/controller/TranslateController.java
@@ -1,16 +1,10 @@
 package com.chalim.backend.controller;
 
-import com.chalim.backend.dto.ResponseDto;
-import com.chalim.backend.apiPayLoad.ApiResponse;
-import com.chalim.backend.dto.TranslateRequest;
-import com.chalim.backend.dto.TranslateResponse;
+import com.chalim.backend.dto.translate.ResponseDto;
+import com.chalim.backend.model.ImageData;
 import com.chalim.backend.service.TranslateService;
-import com.chalim.backend.apiPayLoad.exception.GeneralException;
-import com.chalim.backend.apiPayLoad.code.ErrorReasonDTO;
-import com.chalim.backend.apiPayLoad.code.status.ErrorStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -27,10 +21,11 @@ public class TranslateController {
 
     @PostMapping("/{language}")
     public ResponseEntity<ResponseDto> translateText(@PathVariable String language,
-                                                     @RequestParam("image") MultipartFile imageFile) {
-        String imageName = imageFile.getOriginalFilename();
-        // You can process the image file if needed
-        ResponseDto response = translateService.translateTextData(language, imageName);
+                                                     @ModelAttribute ImageData imageData) {
+        String imageName = imageData.getImageName();
+        MultipartFile imageFile = imageData.getImageFile();
+
+        ResponseDto response = translateService.translateTextData(language, imageName, imageFile);
         return ResponseEntity.ok(response);
     }
 }
