+++ src/main/java/com/chalim/backend/dto/translate/TextData.java
@@ -1,4 +1,4 @@
-package com.chalim.backend.dto;
+package com.chalim.backend.dto.translate;
 
 public class TextData {
     private String transcription;
