+++ src/main/java/com/chalim/backend/dto/translate/TranslateRequest.java
@@ -1,4 +1,4 @@
-package com.chalim.backend.dto;
+package com.chalim.backend.dto.translate;
 
 public class TranslateRequest {
     private String text;
