+++ src/main/java/com/chalim/backend/dto/translate/TranslateResponse.java
@@ -1,4 +1,4 @@
-package com.chalim.backend.dto;
+package com.chalim.backend.dto.translate;
 
 public class TranslateResponse {
     private String translatedText;
